+++ src/Emulator/emulator.py
@@ -1,28 +1,16 @@
 import socket as sock_module
 import sys
+import base
 
 
-class emulator:
+class emulator(base.base):
     def __init__(self, config={}):
-        self.socket = sock_module.socket(
-            sock_module.AF_INET, sock_module.SOCK_DGRAM)
-        self.ip = sock_module.gethostbyname(sock_module.gethostname())
+        base.base(config)
         if config["emulator_ip"] == self.ip:
             print("same ip", self.ip)
             self.socket.bind((config["emulator_ip"], config["emulator_port"]))
         else:
             sys.exit()
-
-    def __del__(self):
-        self.socket.close()
-        del self.socket
-
-    def accept_connections(self):
-        while(1):
-            buffer = self.socket.recv(1024)
-            print(buffer)
-            if buffer == "":
-                break
 
 
 def run(config):
@@ -33,4 +21,4 @@ def run(config):
 
 
 if __name__ == "__main__":
-    run()
+    run(base.test_config)
